Clean up public class variables.
In almost all cases class variables should be private.

This means you hide the underlying data structure.  In this case we
could switch from having 4 different layer groups to 1 without having to
modify all of the underlying classes.

TObjects shouldn't know anything about the world, except for the few
rare instances that the world tells TObject about something, which it
immediately forgets after it is done.